Enhancement - Individually scheduled time for each day in a week
Introduced a new "weekdays" property for each timeperiod. Valid values are 0-6, where 0 is Sunday, 1 is Monday, and so on (same as Date.getDay()). Multiple days can be used with comma as delimiter, and range can also be used with dash. For example, "1-3,5,6" means Monday to Wednesday, plus Friday and Saturday.

Default task settings are updated in this way:
- Weekday schedule starts at 8:00, and previous night ends at 0:45. This is the same as previous default.
- Weekend schedule starts at 9:00, and previous night ends at 2:00.